requirements.txt
streamlit
tensorflow
numpy
Pillow
pandas